Add "pg_use_binary_copy" option
